remove __init__ and testing
