+++ tests/test_30_serializers.py
@@ -12,15 +12,45 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import datetime
+
+import cads_catalogue.database
 import ogc_api_processes_fastapi.models
 
 import cads_processing_api_service.serializers
 
-from . import testing
+
+def get_record(id: str) -> cads_catalogue.database.Resource:
+    return cads_catalogue.database.Resource(
+        resource_id=id,
+        title="ERA5",
+        description="description",
+        abstract="Lorem ipsum dolor",
+        contact=["aaaa", "bbbb"],
+        form="form",
+        citation="",
+        keywords=["label 1", "label 2"],
+        version="2.0.0",
+        variables=["var1", "var2"],
+        providers=["provider 1", "provider 2"],
+        extent=[[-180, 180], [-90, 90]],
+        links=[{"rel": "foo", "href": "http://foo.com"}],
+        documentation="documentation",
+        previewimage="img",
+        publication_date=datetime.datetime.strptime(
+            "2020-01-01T00:00:00Z", "%Y-%m-%dT%H:%M:%SZ"
+        ),
+        record_update=datetime.datetime.strptime(
+            "2020-02-03T00:00:00Z", "%Y-%m-%dT%H:%M:%SZ"
+        ),
+        resource_update=datetime.datetime.strptime(
+            "2020-02-05T00:00:00Z", "%Y-%m-%dT%H:%M:%SZ"
+        ),
+    )
 
 
 def test_process_summary_serializer() -> None:
-    record = testing.get_record("era5-something")
+    record = get_record("era5-something")
     oap_record = cads_processing_api_service.serializers.ProcessSerializer.process_summary_db_to_oap(
         record
     )
@@ -37,7 +67,7 @@ def test_process_summary_serializer() -> None:
 
 
 def test_process_description_serializer() -> None:
-    record = testing.get_record("era5-something")
+    record = get_record("era5-something")
     oap_record = cads_processing_api_service.serializers.ProcessSerializer.process_description_db_to_oap(
         record
     )
